- Extracted types from files containing more than one type.
- Moved contents of TaskExtensions.cs to ResultExtensions.cs, both dealing with IResult.
- Deleted TaskExtensions.cs.

* Note
Nested types still exist in:
- WeakValueDictionary contains ValueCollection as private type (OK).
- LogManager conatins NullLog as private type [Null Pattern] (OK).
- SimpleContainer contains (ContainerEntry and FactoryFactory<T>) as private type (OK).
- EventAggregator contains Handler as private type (OK).
- Conductor partial class contains Collection as partial class which in turn contains (OneActive and AllActive) (Code Smell), We Should use namespaces.

if we were enabling Code Analyzers right now with SDK 7.0.401, we will get CA1052 error among other errors.
change global.json SDK version to 7.0.401 and add these properties to the project:
<PropertyGroup>
  <WarningLevel>7</WarningLevel>
  <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
  <EnforceCodeStyleInBuild>True</EnforceCodeStyleInBuild>
  <EnableNETAnalyzers>True</EnableNETAnalyzers>
  <AnalysisLevel>latest-all</AnalysisLevel>
</PropertyGroup>
