+++ src/Caliburn.Micro.Core/Conductor/Contracts/IConductActiveItem.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Caliburn.Micro
+{
+    /// <summary>
+    /// An <see cref="IConductor"/> that also implements <see cref="IHaveActiveItem"/>.
+    /// </summary>
+    public interface IConductActiveItem : IConductor, IHaveActiveItem
+    {
+    }
+}
