+++ src/Caliburn.Micro.Core/Screen/Contracts/IChild{TParent}.cs
@@ -3,11 +3,12 @@ namespace Caliburn.Micro
     /// <summary>
     /// Denotes a node within a parent/child hierarchy.
     /// </summary>
-    public interface IChild
+    /// <typeparam name="TParent">The type of parent.</typeparam>
+    public interface IChild<TParent> : IChild
     {
         /// <summary>
         /// Gets or Sets the Parent
         /// </summary>
-        object Parent { get; set; }
+        new TParent Parent { get; set; }
     }
 }
